Refactored getGIFtosend to Utils
