+++ botActions/Utils.js
@@ -1,12 +1,15 @@
 const {
     encouragement,
     finalMoments,
+    encouragementGIF,
+    finalMomentsGif
   } = require('./constants');
 
 const getRandomInt = max => {
     return Math.floor(Math.random() * Math.floor(max));
 };
 
+//Create Time Left Message
 module.exports.parseTimeToString = t => {
     let parsedString = '';
     if (t > 0) {
@@ -27,3 +30,19 @@ module.exports.parseTimeToString = t => {
     }
     return parsedString;
   };
+
+//Sets GIF to send
+module.exports.gifToSend = t => {
+    let gifString = '';
+    if (t > 0) {
+        const hours = parseInt(t / 3600000);
+        if (hours > 0) {
+            gifString = encouragementGIF(getRandomInt(4));
+        } else {
+            gifString = finalMomentsGif(getRandomInt(4));
+        }
+    } else {
+        gifString = 'https://media.giphy.com/media/3o7qDEq2bMbcbPRQ2c/giphy.gif';
+    }
+    return gifString;
+};
